+++ Backend/Repositories/IBooksRepository.cs
@@ -9,9 +9,9 @@ namespace OnlineBookShop.Repositories
         public Task<IEnumerable<Book>> GetBooksSortedByPriceAsync(Boolean order);
         public Task<IEnumerable<Book>> GetBooksSearchAsync(string search);
         public Task<Book>? GetBookAsync(int id);
-        public Task<int> GetBookPrice(int id);
+        public Task<float> GetBookPrice(int id);
         public Task<Book> PostBookAsync(Book book);
-        public Task<Book>? PutBookAsync(int id, int newPrice);
+        public Task<Book>? PutBookAsync(int id, float newPrice);
         public Task<Boolean> DeleteBookAsync(int id);
     }
 }
